+++ internal/pkg/runtime/launcher/oci/oci_conmon_linux.go
@@ -34,7 +34,7 @@ type ociError struct {
 }
 
 // Create creates a container from an OCI bundle
-func Create(containerID, bundlePath string) error {
+func Create(containerID, bundlePath string, systemdCgroups bool) error {
 	conmon, err := bin.FindBin("conmon")
 	if err != nil {
 		return err
@@ -116,6 +116,10 @@ func Create(containerID, bundlePath string) error {
 		"--exit-command-arg", containerID,
 	}
 
+	if systemdCgroups {
+		cmdArgs = append(cmdArgs, "--systemd-cgroup")
+	}
+
 	cmd := exec.Command(conmon, cmdArgs...)
 	cmd.Dir = absBundle
 	cmd.Env = append(cmd.Env, fmt.Sprintf("_OCI_SYNCPIPE=%d", 3), fmt.Sprintf("_OCI_STARTPIPE=%d", 4))
@@ -153,7 +157,7 @@ func Create(containerID, bundlePath string) error {
 	// We check for errors from runc (which conmon invokes) via the sync pipe
 	pid, err := readConmonPipeData(syncParent, path.Join(sd, runcLogFile))
 	if err != nil {
-		if err2 := Delete(context.TODO(), containerID); err2 != nil {
+		if err2 := Delete(context.TODO(), containerID, systemdCgroups); err2 != nil {
 			sylog.Errorf("Removing container %s from runtime after creation failed", containerID)
 		}
 		return err